separate the ci test to reduce ci time

+++ test.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 set -e
 echo "" > coverage.txt
+echo $TEST_RACE
 
-GOMAXPROCS=1 go test -tags=embed -timeout 900s `go list ./... | grep -v pdserver`
-GOMAXPROCS=4 go test -tags=embed -timeout 900s -race `go list ./... | grep -v pdserver`
-for d in $(go list ./... | grep -v pdserver | grep -v vendor); do 
-    GOMAXPROCS=4 go test -tags=embed -timeout 900s -race -coverprofile=profile.out -covermode=atomic $d
-    if [ -f profile.out ]; then
-        cat profile.out >> coverage.txt
-        rm profile.out
-    fi
-done
+if [ "$TEST_RACE" = "false" ]; then
+    GOMAXPROCS=1 go test -tags=embed -timeout 900s `go list ./... | grep -v pdserver`
+else
+    GOMAXPROCS=4 go test -tags=embed -timeout 900s -race `go list ./... | grep -v pdserver`
+    for d in $(go list ./... | grep -v pdserver | grep -v vendor); do 
+        GOMAXPROCS=4 go test -tags=embed -timeout 900s -race -coverprofile=profile.out -covermode=atomic $d
+        if [ -f profile.out ]; then
+            cat profile.out >> coverage.txt
+            rm profile.out
+        fi
+    done
+fi
 
 # no tests, but a build is something
 for dir in $(find apps tools -maxdepth 1 -type d) ; do
